backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.5
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.0.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WTF==1.0.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
passlib==1.7.4
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==1.4.2
pyparsing==3.0.9
pytz==2022.7
redis==4.4.0
six==1.16.0
SQLAlchemy==1.4.46
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
